feat: :sparkles: Likes feature

+++ DatingApp/Extensions/ClaimsPrincipleExtensions.cs
@@ -12,9 +12,9 @@ public static class ClaimsPrincipleExtensions
 
         return username;
     }
-    public static int GetUserId(this ClaimsPrincipal user)
+    public static long GetUserId(this ClaimsPrincipal user)
     {
-        var userId = int.Parse(user.FindFirstValue(ClaimTypes.NameIdentifier)
+        var userId = long.Parse(user.FindFirstValue(ClaimTypes.NameIdentifier)
             ?? throw new Exception("Cannot get user ID from token"));
 
         return userId;
